DataVersionsByPartition return value for observable_source_asset (#14354)
## Summary & Motivation

Introduces a way for observable source assets to observe partitions of
an asset.

Usage:
```python
    @observable_source_asset(partitions_def=StaticPartitionsDefinition(["apple", "orange", "kiwi"]))
    def foo():
        return DataVersionsByPartition({"apple": "one", "orange": DataVersion("two")})
```

I'm not convinced this is the right API long-term. I think it's likely
that we want some sort of `Observation` or `ObservationResult` class to
take care of this. But I wanted to get something out that one of our
eager users can work with, and I think this is the lowest-blast radius
API possible.

## How I Tested These Changes