Give pip more hints about which versions of grpcio-tools and grpc-health-checking are valid (#14529)
Summary:
grpcio-tools vX and grpcio-health-checking vX are pinned to grpcio>=X,
so by enforcing a cap on grpcio we are effectively enforcing a cap on
those packages as well. pip should be able to figure that out, but it
appears to be struggling, capping grpcio-health-checking high, and then
telling us that it's an invalid version combination. Add some more hints
to our setup.py to help it out.

Test Plan: BK

## Summary & Motivation

## How I Tested These Changes