Update generation of artificial input names (#14531)
## Summary & Motivation

We sometimes need to create artificial op inputs out of existing op
outputs to make our automatic-subsetting approach work. When this
happens, we were previously generating a name for this input based on
the original output name.

In some rare cases, this automatic subsetting process will happen
multiple times on a single node, which results in us losing track of
what output a key was originally associated with. We could probably fix
this, but instead we can just ignore that original mapping and just
create an arbitrary input name of "artificial_input_N" so that we don't
have to deal with that situation.

## How I Tested These Changes