[ui] Handle network error on repository location reload (#14565)
## Summary & Motivation

When reloading a code location, a network error on the mutation will be
ignored by the UI and the button remains disabled and spinning.

Resolve this by catching the rejected promise on the `client.mutate`
call, which is how the network error is surfaced in direct Apollo API
access like this. (We use `client.mutate` here to avoid needing to use a
React hook to set up the mutation.)

Currently, the rejected promise just falls through and the UI is none
the wiser. By catching the rejection and returning an error payload to
the caller, the UI can handle it and update state accordingly.

## How I Tested These Changes

Add an error to the reload mutation in `schema/roots/mutation.py`, try
to reload the code location. Verify that an error dialog appears with
information about the 500 error.

Repeat with the error removed, verify that reloading the code location
is successful.

<img width="1203" alt="Screenshot 2023-06-01 at 10 18 47 AM"
src="https://github.com/dagster-io/dagster/assets/2823852/2abf4d9c-c756-4734-9a29-ed953fad46d4">