[pythonic resources] Better error message for attempting to use DagsterTypes with Config/ConfigurableResource (#14491)
## Summary

Pydantic will get angry if a user supplies a non-type as an annotation
for a `BaseModel` field (it performs an `issubclass` check so anything
that's not a class causes problems). Users who may be tempted to use
`DagsterTypes` here will run into a Pydantic error message that's not
very fruitful:

```python
class MyAssetConfig(Config):
  my_field: MyDagsterType
```

```python
...
    class MyAssetConfig(Config):
  File "pydantic\main.py", line 204, in pydantic.main.ModelMetaclass.__new__
  File "pydantic\fields.py", line 488, in pydantic.fields.ModelField.infer
  File "pydantic\fields.py", line 419, in pydantic.fields.ModelField.__init__
  File "pydantic\fields.py", line 539, in pydantic.fields.ModelField.prepare
  File "pydantic\fields.py", line 801, in pydantic.fields.ModelField.populate_validators
  File "pydantic\validators.py", line 718, in find_validators
The above exception occurred during handling of the following exception:
TypeError: issubclass() arg 1 must be a class
  File "pydantic\validators.py", line 709, in pydantic.validators.find_validators
```

This PR adds in a new Metaclass for Config with the purpose of catching
this annotation before Pydantic can error. We will give a more useful
error message in this case.

Feels a bit heavy-handed for a single error case, but on the other hand
we're already using a metaclass for ConfigurableResource.

## Test Plan

New unit test.