disable alembic quiet calls for sensor logging test (#14617)
## Summary & Motivation
The stamp alembic call temporarily disables the python logging at the
root logger. This is interfering with some sensor logging tests which
checks to see if some python logs are being captured. This was not an
issue seen before, but is being exposed by the switch to sqlalchemy 2,
which is causing some race conditions between the logging module reset
and the call to log from the sensor.

## How I Tested These Changes
BK, local tests w/ sqlalchemy 2 installed. Should only affect tests.
Will be extra tested with a revert of
https://github.com/dagster-io/dagster/pull/14586