Fix daemon restarting gRPC servers more than expected when running `dagster code-server start`
Summary:
This fixes a latent bit of weirdness in the daemon that was exposed with the addition of the `dagster code-server start` CLI endpoint. The daemon was actually reloading the gRPC servers in two different ways when it was responsible for managing its own gRPC servers:
a) it creates a GrpcServerRegistry on startup and sets a reload_interval there, so the code is automatically reloaded on that interval
b) it calls reload_workspace periodically on an interval as well so tha

What it actually wants to do here is to "refresh" the workspace snapshot - re-generate the snapshots used by the context in case the code has changed, but not necessarily trigger a reload of the code, that is managed elsewhere (by the GrpcServerRegistry if the daemon is managing its own code servers, or by the server itself if it is being managed separately). Before, there wasn't a meaningful distinction between reload and refresh in that second case, but now there is.

Test Plan:
In addition to existing BK coverage on the daemon, verify:
- Run dagster dev with no separate code server, change code, verify that after 1-2 minutes the change is reflected in the daemon (e.g. adding a schedule)
- Run a code server via `dagster api grpc`, make a code change, restart the process, verify after 1-2 minutes the change is reflected in the daemon
- Run a code server via `dagster code-server start`, make a code change, reload via dagit, verify that after 1-2 minutes the change is reflected in the daemon