[ui] Turn on noUncheckedIndexedAccess for TS, Part 1 (#14696)
## Summary & Motivation

The first wave of changes made to allow us to enable
`noUncheckedIndexedAccess` in TypeScript in `packages/core`.

This generally consists of a few different approaches:

- Use the `!` assertion operator on array/object key access. I used this
when it looked like a previous check had validated the presence of the
key, e.g. checking `.length` on an array and subsequently looking up
`[0]`.
- Use `.forEach` instead of `for (let ii = ...)` style loops where
possible, since then we don't need to key into the array.
- Similarly, use `Object.entries()` for loops that currently use
`Object.keys()` and then index into the object based on the key.

It is possible that this will surface new errors, especially in
situations where the `!` assertion is inaccurate. Fortunately, the error
should now be very obvious at that point in code, instead of surfacing
downstream.

There will be one or two more PRs to finish this for OSS.

## How I Tested These Changes

ts, lint, jest.