Handle partitions with invalid parents in auto-materialize (#14706)
Background here:
https://www.notion.so/dagster/Handle-invalid-parents-in-auto-materialize-5fe6dbdfba0f43a6beae88fb61ccc07d?pvs=4

This PR builds upon https://github.com/dagster-io/dagster/pull/14449 and
introduces graceful handling for auto-materialize cases where an invalid
parent exists. Errors are currently raised in
`PartitionMapping.get_upstream_partitions_for_partitions` when mapped
upstream partitions were nonexistent. It is good that an error occurs
when launching backfills and runs in these cases, but undesirable for
the error to occur in auto-materialization. This PR introduces logic to
gracefully handle invalid parents without erroring in auto-materialize.

**Modifies the auto-materialize logic:**
- Only considers partitions that can be materialized. If a partition has
invalid upstreams, it should not be possible to auto-materialize it.
- Modifies `is_reconciled` to recursively check only valid upstream
partitions. If all of a partition's parents are invalid, then returns
`True` if the partition is materialized, `False` if not. This enables
reconciling downstream assets.