feat(dbt): propagate partial parsing to dynamic target (#14762)
## Summary & Motivation
Because we are dynamically setting the target path, dbt does not take
advantage of partial parsing because it is unable to find a previously
parsed artifact in the new target path.

This is unfortunate, because:

1. This artifact already exists, since we require a reference to the
manifest when invoking the CLI.
2. In [dbt's experimental programmatic
API](https://docs.getdbt.com/reference/programmatic-invocations#reusing-objects),
there is an explicit example to reuse pre-constructed objects.

Here, we emulate that behavior. We assume the following:

- Given a dbt project directory (`DBT_PROJECT_DIR`), defined by the user
when instantiating `DbtCli`, we inspect if
`{DBT_PROJECT_DIR}/target/partial_parse.msgpack` exists.
- If this file exists, then copy it over to the dynamic target path
before invoking `DbtCli.cli(...)`. This way, partial parsing is enabled.

## How I Tested These Changes
pytest