feat(dbt-assets): automatically apply metadata replacements using the manifest (#14718)
## Summary & Motivation

The original dbt load functions accepted metadata from two places, the
dbt manifest (how we show column schemas) and from a user supplied
function `node_info_to_definition_metadata_fn`. This latter was
particularly useful in order for partitioned dbt assets to specify a
`partition_expr` metadata attribute for attached io managers.

This PR unblocks this use case for the new asset decorator, but instead
of a helper function, it prefers setting metadata directly on the asset
decorator and then merging that metadata with any found in the dbt
manifest.


## How I Tested These Changes

I tested this on hooli and added unit tests

Co-authored-by: Rex Ledesma <rex@elementl.com>