[ui] Prevent Zoom slider and buttons from closing right panel (#14868)
## Summary & Motivation

Resolves #14861.

Prevent zoom button/slider clicks from propagating to the `SVGViewport`
container, so that they don't get treated as "background" clicks on the
SVG display. This is because "background" clicks will unselect the DAG
nodes, hiding the right panel. Instead, we want to maintain the node
selection even as we zoom in and out.


https://github.com/dagster-io/dagster/assets/2823852/09257616-8c41-44d9-9329-63cdc9a32d19

## How I Tested These Changes

View an asset DAG.

- Click on a node, verify that it is highlighted and shows the right
panel.
- Click off of it on the background of the DAG display, verify that the
panel disappears.
- Click on the node again, then use the zoom buttons and slider to
verify correct zooming behavior, and that the right panel remains open
and the node remains highlighted.