[ui] Refactor error handling in spanTextToSelectionsOrError, show alert on other callsite (#14906)
## Summary & Motivation

This is a fix for the error reported in
https://elementl-workspace.slack.com/archives/C03CCE471E0/p1687272525520469

The `spanTextToSelections` method was designed to throw exceptions when
the text is invalid, but this is a bit of a standard-but-non-ideal case,
and the error should always be handled by the caller. To make it easier
to enforce this in Typescript I converted it to return the Error.

## How I Tested These Changes

I was able to reproduce the error in the report by visiting the
partition page and then changing the URL to reference a partition that
does not exist:


http://localhost:3000/assets/yoyoyoyoyo?view=partitions&default_range=doesnotexist

This would crash the page previously because the selection tries to set
initial state from the URL and fails. I updated this use of
`spanTextToSelections` to display an error the way the other callsites
do. It's admittedly a /bit/ odd to show an error alert from a hook, but
this one just runs the relevant initialize method once so the alert
appears once and not again.

Co-authored-by: bengotow <bgotow@elementl.com>