[auto-materialize][refactor] Use new instance methods (#14917)
## Summary & Motivation

Ok this is a massive set of changes, but please be heartened by the fact
that all tests continue to pass, and there is a net removal of ~100LoC.

High level, we previously treated observable source assets and regular
assets quite differently, leading to us having to fork down different
code paths for each. This also lead to this weird situation where:

- For observable source assets, we factored in data versions, but did
not support partitioned source assets
- For regular assets, we supported partitioned assets, but did not
factor in data versions

Now, we support partitioned observable sources (factoring in data
versions), and the only thing stopping us from factoring in data
versions for regular assets is a single conditional (which I left in
there until we do an evaluation of the perf impact this might incur).

Previously, we had code strewn about in serval different places to deal
with the differences between observable source assets (which require
grabbing asset observation events) and regular assets (where we get
materializations). Now, both go through the same functions (the big ones
being `get_latest_record()` and `get_new_asset_partitions()`). This lets
us avoid a bunch of junk in
`find_parent_materialized_asset_partitions()`.

This will also be hugely more performant in cases where we're evaluating
a large number of partitions of the same asset in a single tick (the
classic example would be unpartitioned -> static partitioned). This
takes advantage of our new `get_latest_materialization_id_by_partition`
functionality.

With a single db call, we can know what the most recent storage id for
each of the static partitions is. Similarly, we'll know the latest
storage id of the unpartitioned asset. To find all the partitions that
have been updated after the unpartitioned asset, we just need to compare
the relevant numbers (so total of 2 db calls). Compare this to the
previous world in which we'd need a separate db call for each partition
to get the latest storage id.


## How I Tested These Changes