feat(dbt): add method to create asset selections from dbt selections (#14980)
## Summary & Motivation
`AssetSelection` is a Dagster specific concept that generalizes dbt
selection syntax. Here, we allow the user to create an `AssetSelection`
using the native dbt selection syntax.

`DbtManifest` is our source of truth for dbt operations, as it wraps the
dbt manifest, the artifact representation of the dbt project. Now, users
can invoke a method on this object to create an `AssetSelection`.

We already have `DbtManifestAssetSelection` representing an
`AssetSelection` in dbt space. We continue to use this object for this
purpose, but we make it simpler.

We keep the following arguments:

| Argument | Notes |
|--------|--------|
| `manifest` | renaming from `manifest_json`, and this is now required
to be a `DbtManifest` |
| `select` | corresponds to dbt `--select` criteria |
| `exclude` | corresponds to dbt `--exclude` criteria |

And remove other arguments:

| Argument | Notes |
|--------|--------|
| `manifest_json_path` | subsumed by `DbtManifest.read(...)` |
| `resource_types` | subsumed under `select` |
| `node_info_to_asset_key` | subsumed by overriding
`DbtManifest.node_info_to_asset_key(...)` |
| `state_path` | deprecated | 


## How I Tested These Changes
pytest