feat(dbt): add method to build schedules from dbt selections (#14997)
## Summary & Motivation

We introduce a `.build_schedule(...)` method on `DbtManifest` to create
a schedule given a selection of dbt assets. This way, users are shielded
from directly interacting with `define_asset_job` and
`ScheduleDefinition`.

The API that we introduce makes the construction of `ScheduleDefinition`
for simple cases. But, we should not seek to recreate a parallel API
that contains all the `ScheduleDefinition` and `define_asset_job`
arguments.

For example, if the user wants to use partition config or yield
`RunRequest` and `SkipReason` in their schedule, then they must use the
underlying Dagster framework API's (`ScheduleDefinition`, `@schedule`,
`define_asset_job`) to accomplish their use cases.

## How I Tested These Changes
pytest