feat(dbt): raise exception by default when dbt command is not successful (#15055)
## Summary & Motivation
The user should have control on what to do when a dbt command fails. By
default, we should expose the failure to the user. Then, the user can
either handle it by:

- suppressing the failure, and setting `raise_on_failure=False`, or
- catching the failure and handling the exception.

Before, the failure was not raised, which caused the underlying
execution in Dagster to not display as failure.

## How I Tested These Changes
pytest