DbtCli.global_config -> global_config_flags (#15067)
## Summary & Motivation

I'm aiming to remove the experimental label from DbtCli, which will
allow me to start switching over the tutorial. As a precursor to this, I
did a final review of the API. The only thing that came up was that I
found the name `global_config` to be a tiny bit confusing - I thought it
make accept a dictionary of config key value pairs, when in fact it
accepts a list of command line flags. If I understand correctly from the
[dbt
documentation](https://docs.getdbt.com/reference/global-configs/command-line-flags),
flags are one way to set dbt global config, and this particular API sets
flags.

## How I Tested These Changes