Avoid creation of local directories by the GCS IO Manager
Prior to this change the underlying behaviour from the base
`UPathIOManager` would be invoked when the user had configured the
use of a GCS IO Manager. This would mostly result in the (perhaps
innocuous) creation of local directories, however when running on
a filesystem where the user does not have write access it would
cause IOErrors that would cause job runs to fail.

This change copies the approach taken in the AWS S3 IO Manager, as
(just like S3) GCS doesn't really have the concept of directories
and there is no need to try to create them (either locally or in
the remote store).