Render @deprecated in API docs (#15077)
## Summary & Motivation

Closes #15003 

Throughout our code we use `@deprecated` as a decorator on deprecated
classes/methods. This was not being rendered in our API docs.

This PR extends our custom sphinx extension to render `@deprecated` as a
special badge at the top of the API doc entry. It works as follows:

- Use a docstring processing hook to inject `.. deprecated::
<placeholder>` at the start of the docstring for any object having a
`@deprecated` tag. This is a builtin sphinx deprecated directive that is
ultimately rendered like this:

```
<div class="deprecated">
  <p>
    <span class="deprecated">
      Deprecated in version <placeholder>.
    </span>
  </p>
</div>
```

- Because there are a number of existing `@deprecated` in the codebase
and it is not clear at what version the target was deprecated, I used a
placeholder here that we remove later. The placeholder is necessary
because `.. deprecated::` requires an argument.
- After the sphinx doctree is built, look up all `versionmodified` nodes
(the docutils node type corresponding to sphinx `deprecated` directive)
and replace the contained text with just `Deprecated`.
- The result in the final HTML:

```
<div class="deprecated">
  <p>
    <span class="deprecated">
      Deprecated
    </span>
  </p>
</div>
```

I added CSS to the NextJS site for `span.deprecated` to style this as a
badge. I am sure it doesn't quite match what we want here, so cc
@bengotow for CSS tweaks:

<img width="400" alt="image"
src="https://github.com/dagster-io/dagster/assets/1531373/7a8bdd2f-de0b-4498-acb9-3fc73a2bded1">

## How I Tested These Changes

Rendered the docs.