[docs] Dark launch Swiftype search (#15111)
## Summary & Motivation

Begin testing Swiftype search by dark-launching queries, debounced in
parallel to the existing search dialog UI. This should allow us to
observe search quality without displaying anything new in the UI.

## How I Tested These Changes

In local development and preview deployment, perform a search via the
search dialog. Verify that the query is debounced and sent to the
`/api/search` serverless function, which proxies to Swiftype. Verify
that the resulting records look good in the Network tab, and have no
effect on the current UI.