Use typing.get_type_hints for input/output and resource/config type inference (#14707)
## Summary & Motivation

Fixes #14571 

Currently, `Parameter.annotation` is used when inferring the dagster
type from an op compute function signature. This causes an error if
either: (a) a string is used directly by the user for the parameter
annotation; (b) `from __future__ import annotations` is used. This is
because both cases cause the value of `Parameter.annotation` to be a
string, and our dagster type resolution logic does not handle strings.

This PR replaces direct access to `Parameter.annotation` with
`typing.get_type_hints`, which handles evaluation of a string annotation
into an actual class which can be handled by our dagster type resolution
logic. Note that this only works if the referents of a string annotation
is module-level scope-- it will fail for a local scope.

Still this is an upgrade and will fix the problem in most cases. It
will:

- Allows users to use string annotations or `from __future__ import
annotations` together with dagster type signature inference
- Allows us to use `from __future__ import annotations` in our own code,
which is necessary to allow Ruff to auto-sort typing-only imports-- see
https://github.com/dagster-io/dagster/pull/14675. Not clear whether this
is a good idea.

## How I Tested These Changes

New unit test using string annotations.